More work on siap, particularly on the data model.
Tuples now serialize to arrays (need to check if that's a problem for
querulator, but probably not, it uses lists for SQL sets).

A test data set now ingests with what I believe are all parameters necessary
of a complete siap answer.

In consequence, we now have SQL array types to worry about.  typesystems
and the (still horribly named) parsing.typeconversions now handle them
more or less.


git-svn-id: http://svn.ari.uni-heidelberg.de/svn/gavo/python/trunk@111 ca62dbe9-ed84-400c-aab7-804ce7e0b8cc