The default html table now generates title lines every so often.  Also, it returns a deferred when doing this, thus giving the reactor a chance to do other things while rendering huge tables.
New MAKE_WITH_NULL macro for import boosters.

You can now block invdividual services while you do maintanence on them.
For now, this is implemented as a "blocked" file alongside to the rd and
a cheap attribute on the rd itself.  We should think of a better
solution some time.

Finally removed booster attribute on ColumnGrammars.


git-svn-id: http://svn.ari.uni-heidelberg.de/svn/gavo/python/trunk@195 ca62dbe9-ed84-400c-aab7-804ce7e0b8cc