This is a release likely to have broken quite a few things, since there are changes deep within DbBasedCore.  The tests ran fine...
You can now write custom renderers in python in OutputFields.

OutputFields have a new attribute select to override the default
contribution of an output field to a SELECT list (its dest, that still
is the default for get_select()).

Parameter names starting with an underscore are now not displayed in
the parameters summary in HTML responses.

DbBasedQuery now correctly operates on OutputFields (rather than plain
DataFields) in its output definition.

The whole source="None" business in OutputFields introduced recently was
a pain.  Removed it again, adjusted simbad formatter accordingly.



git-svn-id: http://svn.ari.uni-heidelberg.de/svn/gavo/python/trunk@237 ca62dbe9-ed84-400c-aab7-804ce7e0b8cc