Resourcebased web pages now have a sidebar containing meta information and other stuff.
Gavoimp now leaves a timestamp in stateDir after an import.
importparser.getRd gets this timestamp if present and adds it as a
dateUpdated meta item (as a ISO date string).

Added a hack to push in an http forwarder's x-forwarded-host header into the
request when present to keep query forms from going off to the weird
server address I currently have.  I'd love to have a way of finding
out the port on the forwarder, but alas, I can't see one, so I hardcode
80.

New module parsing.elgen to register "element generators", another
attempt to give some kind of macro facility to resource descriptors.
The first application is the siapOutput elgen that defines the SIAP
output fields.

Added a user help file (to have something to link to from the sidebar:-)

MetaItems now mix in MetaMixin.  This means that they could have internal
structure.  Right now, this *cannot* be used for structured meta items
(like contributors or the like), since I don't have a good way to let
the registry code handle this.  It's mainly intended to enrich meta
items for HTML rendering, but we probably should go all the way to
hierarchical meta interformation.


git-svn-id: http://svn.ari.uni-heidelberg.de/svn/gavo/python/trunk@247 ca62dbe9-ed84-400c-aab7-804ce7e0b8cc