Great Refactor, part II (actually, just preparations for easier handling of the ADQL public interface and the related privilege issues; I reckon we'll have another 20 of such sessions until I start being happy).
TableWriters are now constructed with complete TableDefs rather than a
name and fields.  Thus, we can transport richer information in the
future 

Consequently, did away with the horrible hack of passing the schema to
feed; this means we can have shared tables in schemas -- yay!

TableDefs now receive their resource descriptor on construction.  It
is allowed to build "free" TableDefs that have an RD None, but you
cannot use these to write DB tables.

Finally added a descriptor for fielddescriptions in dc_tables.  However,
it's currently only used in sqlsupport.

Changed lots of unit tests to use the test.vord RD to make them work
with the changed TableWriter.



git-svn-id: http://svn.ari.uni-heidelberg.de/svn/gavo/python/trunk@407 ca62dbe9-ed84-400c-aab7-804ce7e0b8cc