You can now pass profile names to SimpleQuerier.
getDbConnection uses the default profile when you pass it profile=None

Fixed a bug in which in tar files, non-disambiguated names would be in
the root of the archive.

Added a getWaitingChild method to nodebuilder to locate siblings (or
ancestors of siblings) in the tree being built.  Used that to become
independent of the currentService property in srvInput and srvOutput.

Made product table a shared table again (oops...).

Fixed a nasty bug where DataFieldLists as children of TableDefs were not
copied by the copy method.  It's high time I overhauled those horrible
"Records".



git-svn-id: http://svn.ari.uni-heidelberg.de/svn/gavo/python/trunk@414 ca62dbe9-ed84-400c-aab7-804ce7e0b8cc