Moved mkrd to user and ported it.
Structures now have a attrSeq attribute containing attribute, but,
contrary to managedAttrs.values(), each one only once.  This obviates
the need for all the horrible set(..value()) constructs.

Added utils.algotricks for somewhat trickier algorithms and put a
simple indexed graph class and a function for a topological sort
in there.

AttributeDefs can now have a "before" attribute naming an attribute
the current one should stand before in the attrSeq.  This probably
is most interesting for later serialization.  Added this to rscdesc
(where it matters for id resolution) and column (where it's just
pretty).

New function structToETree in mkrd; this allows serialization of
Structures into XML.  Maybe I'll promote this to base if it turns out
it's useful.

New class EventStorage in mkrd; this allows recording of parse
events and later replay.  This is another thing that should be useful
other contexts.

Fixed StructAttribute.iterEvents to cope with polymorphous attributes.

iterEvents now checks if an attribute has an iterEvents method rather
than if it's derived from AtomicAttribute to decide wheter to unparse
it.  This allows CollOfAtomsAttributes to be copied.

The products table now contains mime types, and the defineProducts
rowgen has an additional argument to define it.



git-svn-id: http://svn.ari.uni-heidelberg.de/svn/gavo/python/trunk@600 ca62dbe9-ed84-400c-aab7-804ce7e0b8cc