Ditched the longdescr and longmime attributes on columns (they sucked).
Added a note attribute on columns and taught the HTML table renderer to
do something with it.

Added notes meta and taught the tableinfo renderer to do something with
it.

Changed tableinfo URI scheme -- the table name is now in the URL rather
than a query parameter.

New tablenote renderer for note metadata.

Changes in htmltable and the meta templates to accomodate table notes.

First steps towards a sanitation of the whole web exception processing
code.

NOTE: You need to gavoimp --system __system__/dc_tables && gavopublish -Ma
after this




git-svn-id: http://svn.ari.uni-heidelberg.de/svn/gavo/python/trunk@873 ca62dbe9-ed84-400c-aab7-804ce7e0b8cc