New "gavo init" command to set up (or fix) the file system hierarchy.
Better formatting of error messages in errhandle -- paragraphs are
now recognized, and lines with whitespace in front are left alone.

Better error messages from base.getDBConnection.

New config item for gavo group (needs to be fleshed out; the log files should
all have this group and be writable by it).



git-svn-id: http://svn.ari.uni-heidelberg.de/svn/gavo/python/trunk@1373 ca62dbe9-ed84-400c-aab7-804ce7e0b8cc