Killing resourcebased, part II:
Ex-ServiceResource is not a resource any more but a _formatOutput
function with some ancillary data.  The stuff moved to a new
web.serviceresults module.

Former svc.customwidgets.OutputFormat moved to serviceresults as well
and discovers the output formats available dynamically.

Results of utils.buildClassResolver now have the registry dict
available as the registry attribute.

Removed voplotEnable.  VOPlot output is now available if voplotCodeBase
is set.


git-svn-id: http://svn.ari.uni-heidelberg.de/svn/gavo/python/trunk@1474 ca62dbe9-ed84-400c-aab7-804ce7e0b8cc