BROKEN
TableBasedCores now adapt themselves to renderers; this lets us keep
having vizierexprs by default in form renderers without all the
horrible auto-mapping logic we used to have for InputKeys, plus it lets
us do PQL-like CondDescs when that is called for.

Renderers now have a parameterStyle class attribute that says how they
would like automatic condDescs to perform.

Moved most of vizierexprs from base to protocols.  Some central parts
remained in base as sqlmunge.

Scrapped the horrible formalType of InputKeys.

FormMixin is now in formrender, together with the inputKey-to-formal
logic that previously was part of InputKey.



git-svn-id: http://svn.ari.uni-heidelberg.de/svn/gavo/python/trunk@1629 ca62dbe9-ed84-400c-aab7-804ce7e0b8cc