Cores' output tables are now "prototypes".  You can add to them using the <outputTable> element.
The ssap core is the first to provide such a prototype (scs and siap
should certainly follow).

New fromTable attribute on OutputTables; this now defaults to the
queried table on cores, and it is used for the verbLevel and autocols
attributes.  In addition, params are now also copied (though this
doesn't really work for autocols yet since you cannot reference params
in ColRefListAttributes so far).



git-svn-id: http://svn.ari.uni-heidelberg.de/svn/gavo/python/trunk@1643 ca62dbe9-ed84-400c-aab7-804ce7e0b8cc