The UWS jobs table now has a default timeout of 60 seconds (rather than none as before).
Null transitions in UWS are now ignored.

Did away with the uwsresults table.  It was overkill just of the purpose
of storing the result mime.  Ideally, after installing this release,
you'd wait for an empty UWS and then say gavo imp --system //uws.
Anyway, results are now stored in a pickle in the jobs dir.




git-svn-id: http://svn.ari.uni-heidelberg.de/svn/gavo/python/trunk@1805 ca62dbe9-ed84-400c-aab7-804ce7e0b8cc