Creator.name-s are now parsed to create separate creator metas.
This was a welcome opportunity to sanitise the generation of special
meta items.  This is now done in addMeta with the help of a new method
doMetaOverride

The horrible makeMetaValue and its ilk is still there, to be removed
really soon now.

Also: Allowing digits in the tails of meta key atoms.



git-svn-id: http://svn.ari.uni-heidelberg.de/svn/gavo/python/trunk@4744 ca62dbe9-ed84-400c-aab7-804ce7e0b8cc